libusbmuxd: use pthread_cancel to make usbmuxd_unsubscribe work properly
This actually only affects inotify-enabled builds. Since the faulty
SIGINT pthread 'killing' was replaced with a proper solution, this
fix is required for the situation where usbmuxd isn't initially
running or was terminated; in these cases libusbmuxd was hanging
inside a read() system call (waiting for an inotify event) causing
pthread_join in usbmuxd_unsubscribe() to wait infinitely.