udev: Set "deactivated" USB configuration on iOS devices by default
With the introduction of iOS 7, devices should start of in the "deactivated"
USB configuration "0" by default to not trigger a trust dialog.
Once the devices are identified, usbmuxd will set the correct USB configuration
itself and immediately run the preflight worker code to ensure a trust
relationship with the host is established. This change fixes the trust dialog
from appearing during hotplug of paired devices. The last remaining issue is
that current kernel code still sets the USB configuration to "1" by default
before the udev rules and thus causes multiple connection beeps.