Fix support for more than one device with systemd by extending usbmuxd usage
Upon connecting a new device, usbmuxd was not notified to scan for it if
systemd was used as the activation method. This change introduces the
"--disable-hotplug" and "--enable-exit" options to have a finer control
over usbmuxd's signaling behavior. The convenience option "--systemd"
complements the "--udev" option and setups the right settings for systemd.