log: Fix timestamps being printed incorrectly when running in foreground
Due to usage of wrong function (get_tick_count) the timestamps have
been printed incorrectly based on clock_gettime. This commit fixes it by
using gettimeofday correctly and also makes sure that this is thread-safe
by using localtime_r if available.
Furthermore, this commit will also have the effect that when logging
through syslog we don't determine the current time anymore because the
timestamp is not even used.