main: Increase listen socket backlog size
In environments with a larger number of devices, especially when these
are connected at the time usbmuxd starts, there will be a lot of simultaneous
connection attemps. With a backlog size of 5 these connection attempts will
easily get a ECONNREFUSED thus failing to perform the required preflight
operations. Increasing this to 256 will help to mitigate this.