usb: Allow configuration 5 after verifying it is actualy usable
VMware on macOS somehow exposes a bad configuration 5 for iDevices. Trying
to use it breaks things and can end up in a kernel panic on the device. The
code change introduced with this commit tries its best to make sure the USB
configuration 5 is not 'bad' before switching to it, and otherwise falling
back to configuration 4.