Fix preflight for older devices
On older devices with iOS 5 and even before there is no "ProductName", only
"ProductType" or "DeviceClass" (which is still present).
usbmuxd fails to connect these devices, because it can't receive product name.

"DeviceClass", like "ProductVersion", can be retrieved even in locked state,
so this commit changes it to use that instead.