Implement OpenRC support by providing RC service script
This commit adds a service script for an OpenRC init system.
It intends to mitigate a "race" caused by a start of long-living process
(that is `usbmuxd` itself) by udev rules, which is highly discouraged by
the `udev` manpage:
"Starting daemons or other long-running processes is not allowed; the
forked processes, detached or not, will be unconditionally killed after
the event handling has finished."
There is a bug in `usbmuxd`, which causes the hang, but this is to be
handled in a different commit.

This adds two new `configure` flags:
* `--with-openrc`: toggles on or off (default - off) OpenRC service
  installation and udev activation via that. Conflicts with systemd, so
  the latter must be disabled explicitly:
  `--with-openrc --without-systemd`
* `--with-rcservicedir=DIR`: specifies directory where to install service
  file. Defaults to `$sysconfdir/init.d` (which is `/etc/init.d`)

Issue: https://github.com/libimobiledevice/usbmuxd/issues/210
Signed-off-by: BalkanMadman <zurabid2016@gmail.com>