configure.ac: Do not use '+=' to set CFLAGS to be POSIX-compliant
The '+=' operator used in configure.ac to append to the CFLAGS variable
is present in Bash, but not the POSIX sh specification.
Therefore, the aforementioned part of the configure.ac (from which the
configure script is obtained) might not run correctly under non
Bash-like POSIX compliant shells (dash - default shell on Debian, ash,
etc).

Bug: https://bugs.gentoo.org/924200
Signed-off-by: Zurab Kvachadze <zurabid2016@gmail.com>