usb: check if the device is alive before remove the device
The device may disconnect and connected (hotplug) before the
initialization is finished. When disconnected, the usbmuxd will try to
call device_remove in usb.c. However, the device is added to devicelist(in device.c)
after initialization. Thus it cannot find the corresponding usb device
and unable to remove the device.